+++ src/darwin/uikit/text.m
@@ -8,7 +8,7 @@
 @end
 
 int
-fx_text_init (fx_t *app, const char *value, size_t len, double x, double y, double width, double height, fx_text_t **result) {
+fx_text_init (fx_t *app, const char *value, size_t len, float x, float y, float width, float height, fx_text_t **result) {
   FXText *native_text = [[FXText alloc] initWithFrame:CGRectMake(x, y, width, height)];
 
   native_text.text = [[NSString alloc] initWithBytes:value length:len encoding:NSUTF8StringEncoding];
